canvas-agent-grader/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
google-genai==0.3.0
pandas==2.2.3
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==11.1.0
beautifulsoup4==4.12.3
